fix: move getSkillBinsMap to skill-bins.ts, remove skills/ import from exec
bash-tools.exec.ts now imports getSkillBinsMap and matchSkillByCommand
directly from ./skill-bins.ts, avoiding the architecture boundary violation
where exec tools shouldn't import from skills/.
